refactor(ext): move popup components to sidepanel/ structure
Reorganizes the extension UI from popup/ into sidepanel/ split by responsibility:
- sidepanel/views/ - full-screen views (MainView, AnalysisView, ResultsView, ChatView, LoginView)
- sidepanel/components/ - smaller widgets (PromoBanner)
- sidepanel/shared/ - visual primitives (Icons, DeepSightSpinner, MicroDoodleBackground, SynthesisView, FeatureCTAGrid, doodles/)

Also renames ChatDrawer -> ChatView (no more 'drawer' concept in the sidebar).
src/popup.tsx -> src/sidepanel/index.tsx (webpack entry will be updated in next task).

All relative imports adjusted. Typecheck clean. Pre-existing test failures
(Popup.test.tsx) unchanged - that test will be renamed and fixed in PR 2 final task.
