+++ extension/src/sidepanel/index.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { createRoot } from "react-dom/client";
-import { App } from "./popup/App";
+import { App } from "./App";
 
 const container = document.getElementById("root");
 if (container) {
