+++ extension/src/sidepanel/shared/DeepSightSpinner.tsx
