+++ extension/src/sidepanel/shared/FeatureCTAGrid.tsx
