+++ extension/src/sidepanel/shared/Icons.tsx
