+++ extension/src/sidepanel/shared/MicroDoodleBackground.tsx
