+++ extension/src/sidepanel/shared/SynthesisView.tsx
