+++ extension/src/sidepanel/shared/doodles/DoodleIcon.tsx
