+++ extension/src/sidepanel/shared/doodles/doodlePaths.ts
